Replace smithsonian script with the refactored version
