+++ openverse_catalog/dags/providers/provider_workflows.py
@@ -11,9 +11,7 @@ from providers.provider_api_scripts.metropolitan_museum import MetMuseumDataInge
 from providers.provider_api_scripts.museum_victoria import VictoriaDataIngester
 from providers.provider_api_scripts.provider_data_ingester import ProviderDataIngester
 from providers.provider_api_scripts.science_museum import ScienceMuseumDataIngester
-from providers.provider_api_scripts.smithsonian_refactored import (
-    SmithsonianDataIngester,
-)
+from providers.provider_api_scripts.smithsonian import SmithsonianDataIngester
 from providers.provider_api_scripts.smk import SmkDataIngester
 from providers.provider_api_scripts.stocksnap import StockSnapDataIngester
 from providers.provider_api_scripts.wikimedia_commons import (
@@ -193,12 +191,6 @@ PROVIDER_WORKFLOWS = [
     ),
     ProviderWorkflow(
         provider_script="smithsonian",
-        start_date=datetime(2020, 1, 1),
-        schedule_string="@weekly",
-        load_timeout=timedelta(hours=4),
-    ),
-    ProviderWorkflow(
-        provider_script="smithsonian_refactored",
         ingestion_callable=SmithsonianDataIngester,
         start_date=datetime(2020, 1, 1),
         schedule_string="@weekly",
